refactor(desktop): 提取 dream collector 与提交信息编排
